+++ src/onprem/celery.py
@@ -62,6 +62,11 @@ def prepare_beat_schedule() -> dict[str, dict]:
             'schedule': Env.CELERY_PROCESS_INTERVAL_REPORTS_SCHEDULE,
             'args': (),
         },
+        'remove_old_shards': {
+            'task': 'onprem.tasks.remove_old_shards',
+            'schedule': Env.CELERY_REMOVE_OLD_SHARDS_SCHEDULE,
+            'args': (Env.CELERY_REMOVE_OLD_SHARDS_DAYS.as_int(), ),
+        },
     }
     disabled = []
     for name, inner in schedule.items():
@@ -119,6 +124,7 @@ app.conf.task_routes = {
     'onprem.tasks.collect_metrics': {'queue': 'c-scheduled', 'priority': 0},
     'onprem.tasks.collect_resources': {'queue': 'c-scheduled', 'priority': 0},
     'onprem.tasks.delete_expired_metrics': {'queue': 'c-scheduled', 'priority': 0},
+    'onprem.tasks.remove_old_shards': {'queue': 'c-scheduled'},
 }
 app.conf.timezone = Env.CELERY_TIMEZONE.as_str()
 app.conf.broker_connection_retry_on_startup = True